implement toggle all
