+++ src/App.tsx
@@ -1,17 +1,20 @@
 /* eslint-disable jsx-a11y/control-has-associated-label */
-import React, { useContext, useState } from 'react';
-import { TodoItem } from './components/TodoItem';
-import { DispatchContext, TodosContext } from './state/State';
+import React, { ChangeEvent, useContext, useState } from 'react';
+import { DispatchContext } from './state/State';
+import { TodoList } from './components/TodoList';
 
 export const App: React.FC = () => {
   const [value, setValue] = useState('');
-  const { todos } = useContext(TodosContext);
   const dispatch = useContext(DispatchContext);
 
   const handleOnSubmit = (event: React.FormEvent) => {
     event.preventDefault();
     dispatch({ type: 'addTodo', payload: value });
     setValue('');
+  };
+
+  const handleToggleAll = (event: ChangeEvent<HTMLInputElement>) => {
+    dispatch({ type: 'toggleAll', payload: event.target.checked });
   };
 
   return (
@@ -37,50 +40,11 @@ export const App: React.FC = () => {
           id="toggle-all"
           className="toggle-all"
           data-cy="toggleAll"
+          onChange={handleToggleAll}
         />
         <label htmlFor="toggle-all">Mark all as complete</label>
 
-        <ul className="todo-list" data-cy="todoList">
-          {todos.map(todo => (
-            <TodoItem key={todo.id} todo={todo} />
-          ))}
-
-          {/* <li>
-            <div className="view">
-              <input type="checkbox" className="toggle" id="toggle-view" />
-              <label htmlFor="toggle-view">asdfghj</label>
-              <button type="button" className="destroy" data-cy="deleteTodo" />
-            </div>
-            <input type="text" className="edit" />
-          </li>
-
-          <li className="completed">
-            <div className="view">
-              <input type="checkbox" className="toggle" id="toggle-completed" />
-              <label htmlFor="toggle-completed">qwertyuio</label>
-              <button type="button" className="destroy" data-cy="deleteTodo" />
-            </div>
-            <input type="text" className="edit" />
-          </li>
-
-          <li className="editing">
-            <div className="view">
-              <input type="checkbox" className="toggle" id="toggle-editing" />
-              <label htmlFor="toggle-editing">zxcvbnm</label>
-              <button type="button" className="destroy" data-cy="deleteTodo" />
-            </div>
-            <input type="text" className="edit" />
-          </li>
-
-          <li>
-            <div className="view">
-              <input type="checkbox" className="toggle" id="toggle-view2" />
-              <label htmlFor="toggle-view2">1234567890</label>
-              <button type="button" className="destroy" data-cy="deleteTodo" />
-            </div>
-            <input type="text" className="edit" />
-          </li> */}
-        </ul>
+        <TodoList />
       </section>
 
       <footer className="footer">
